seperate to do_set_attribute
