+++ frame/uniques/src/lib.rs
@@ -1042,40 +1042,7 @@ pub mod pallet {
 				.map(|_| None)
 				.or_else(|origin| ensure_signed(origin).map(Some))?;
 
-			let mut collection_details =
-				Collection::<T, I>::get(&collection).ok_or(Error::<T, I>::UnknownCollection)?;
-			if let Some(check_owner) = &maybe_check_owner {
-				ensure!(check_owner == &collection_details.owner, Error::<T, I>::NoPermission);
-			}
-			let maybe_is_frozen = match maybe_item {
-				None => CollectionMetadataOf::<T, I>::get(collection).map(|v| v.is_frozen),
-				Some(item) => ItemMetadataOf::<T, I>::get(collection, item).map(|v| v.is_frozen),
-			};
-			ensure!(!maybe_is_frozen.unwrap_or(false), Error::<T, I>::Frozen);
-
-			let attribute = Attribute::<T, I>::get((collection, maybe_item, &key));
-			if attribute.is_none() {
-				collection_details.attributes.saturating_inc();
-			}
-			let old_deposit = attribute.map_or(Zero::zero(), |m| m.1);
-			collection_details.total_deposit.saturating_reduce(old_deposit);
-			let mut deposit = Zero::zero();
-			if !collection_details.free_holding && maybe_check_owner.is_some() {
-				deposit = T::DepositPerByte::get()
-					.saturating_mul(((key.len() + value.len()) as u32).into())
-					.saturating_add(T::AttributeDepositBase::get());
-			}
-			collection_details.total_deposit.saturating_accrue(deposit);
-			if deposit > old_deposit {
-				T::Currency::reserve(&collection_details.owner, deposit - old_deposit)?;
-			} else if deposit < old_deposit {
-				T::Currency::unreserve(&collection_details.owner, old_deposit - deposit);
-			}
-
-			Attribute::<T, I>::insert((&collection, maybe_item, &key), (&value, deposit));
-			Collection::<T, I>::insert(collection, &collection_details);
-			Self::deposit_event(Event::AttributeSet { collection, maybe_item, key, value });
-			Ok(())
+			Self::do_set_attribute(maybe_check_owner, collection, maybe_item, key, value)
 		}
 
 		/// Clear an attribute for a collection or item.
